+++ packages/babel-plugin-react-intl/src/options.ts
@@ -9,8 +9,9 @@ export interface OptionsSchema {
   moduleSourceName?: string;
   extractSourceLocation?: boolean;
   messagesDir?: string;
-  overrideIdFn?: (id: string, defaultMessage: string, descriptor: string) => string;
+  overrideIdFn?: (id: string, defaultMessage: string, descriptor: string, filePath: string) => string;
   removeDefaultMessage?: boolean;
   extractFromFormatMessageCall?: boolean;
   additionalComponentNames?: string[];
+  outputEmptyJson?: boolean;
 }