+++ packages/babel-plugin-react-intl/test/fixtures/overrideIdFn/actual.js
@@ -1,9 +1,5 @@
 import React, {Component} from 'react';
-import {
-  defineMessages,
-  FormattedMessage,
-  FormattedHTMLMessage,
-} from 'react-intl';
+import {defineMessages, FormattedMessage} from 'react-intl';
 
 const msgs = defineMessages({
   header: {
@@ -39,10 +35,12 @@ export default class Foo extends Component {
             metadata: 'Additional metadata content.',
           }}
         />
-        <FormattedHTMLMessage
-          id="foo.bar.delta"
-          defaultMessage="<h1>Hello World!</h1>"
-          description="The default message."
+        <FormattedMessage
+          defaultMessage="NO ID"
+          description={{
+            text: 'Something for the translator. Another description',
+            metadata: 'Additional metadata content.',
+          }}
         />
       </div>
     );